profile DB

+++ db/migrate/20190908053308_create_trainings.rb
@@ -2,7 +2,7 @@ class CreateTrainings < ActiveRecord::Migration[5.2]
   def change
     create_table :trainings do |t|
       t.string :name
-      t.integer :MET
+      t.float :MET
       t.integer :time
       t.integer :kcal
     end
